fix unit tests with new correlation plot
